+++ src/database/tests/DatabaseConnectionTest.php
@@ -25,6 +25,7 @@ use Hyperf\Database\Query\Grammars\Grammar;
 use Hyperf\Database\Query\Grammars\Grammar as QueryGrammar;
 use Hyperf\Database\Query\Processors\Processor;
 use Hyperf\Database\Schema\Builder;
+use HyperfTest\Database\Stubs\ExceptionPDO;
 use Mockery as m;
 use PDO;
 use PDOException;
@@ -376,9 +377,11 @@ class DatabaseConnectionTest extends TestCase
         $mock = $this->getMockConnection(['tryAgainIfCausedByLostConnection'], $pdo);
         $mock->expects($this->once())->method('tryAgainIfCausedByLostConnection');
 
-        $method->invokeArgs($mock, ['', [], function () {
-            throw new QueryException('', [], new Exception());
-        }]);
+        $method->invokeArgs($mock, [
+            '', [], function () {
+                throw new QueryException('', [], new Exception());
+            },
+        ]);
     }
 
     public function testRunMethodNeverRetriesIfWithinTransaction()
@@ -395,9 +398,11 @@ class DatabaseConnectionTest extends TestCase
         $mock->expects($this->never())->method('tryAgainIfCausedByLostConnection');
         $mock->beginTransaction();
 
-        $method->invokeArgs($mock, ['', [], function () {
-            throw new QueryException('', [], new Exception());
-        }]);
+        $method->invokeArgs($mock, [
+            '', [], function () {
+                throw new QueryException('', [], new Exception());
+            },
+        ]);
     }
 
     public function testFromCreatesNewQueryBuilder()
@@ -450,6 +455,24 @@ class DatabaseConnectionTest extends TestCase
         $schema = $connection->getSchemaBuilder();
         $this->assertInstanceOf(Builder::class, $schema);
         $this->assertSame($connection, $schema->getConnection());
+    }
+
+    public function testThrowExceptionWhenPDODestruct()
+    {
+        $connection = $this->getMockConnection(pdo: new ExceptionPDO(true));
+
+        $connection->setReadPdo(new ExceptionPDO(true));
+
+        $connection->disconnect();
+
+        $this->assertNull($connection->getPdo());
+        $this->assertNull($connection->getReadPdo());
+
+        $connection = $this->getMockConnection(pdo: new ExceptionPDO(true));
+
+        $connection->setPdo($pdo2 = new ExceptionPDO(false));
+
+        $this->assertSame($pdo2, $connection->getPdo());
     }
 
     protected function getMockConnection($methods = [], $pdo = null)