+++ src/metric/src/Listener/OnWorkerStart.php
@@ -123,7 +123,7 @@ class OnWorkerStart implements ListenerInterface
         );
 
         $timerInterval = $this->config->get('metric.default_metric_interval', 5);
-        $timerId = $this->timer->tick($timerInterval * 1000, function () use ($metrics) {
+        $timerId = $this->timer->tick($timerInterval, function () use ($metrics) {
             $server = $this->container->get(\Swoole\Server::class);
             $serverStats = $server->stats();
             $this->trySet('gc_', $metrics, gc_status());